Refactor serialiser.py into Gen_Loader.py and Loader.cpp.dm
