+++ Sprocket/GenerateComponents.py
@@ -2,7 +2,7 @@ import json
 import os
 
 from Datamatic import definitions
-from Datamatic import serialiser
+from Datamatic import Gen_Loader
 from Datamatic import lua_backend
 from Datamatic import lua_bindings
 from Datamatic import anvil_inspector
@@ -20,9 +20,9 @@ inspector = os.path.join(sprocket_base, "Anvil", "Panels", "Inspector.cpp")
 with open("ComponentSpec.json") as specfile:
     spec = json.loads(specfile.read())
 
-definitions.generate(spec, components_h)
-serialiser.generate(spec, serialiser_cpp)
-lua_backend.generate_header(spec, lua_h)
-lua_backend.generate_cpp(spec, lua_cpp)
-lua_bindings.generate(spec, lua)
-anvil_inspector.generate(spec, inspector)
+#definitions.generate(spec, components_h)
+Gen_Loader.generate(spec, serialiser_cpp)
+#lua_backend.generate_header(spec, lua_h)
+#lua_backend.generate_cpp(spec, lua_cpp)
+#lua_bindings.generate(spec, lua)
+#anvil_inspector.generate(spec, inspector)
